fix: Problem logging out when refreshing
